pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "face-liveness"
version = "1.0.0"
description = "Face Liveness Detection — heuristic + deep learning anti-spoofing"
requires-python = ">=3.11,<3.12"
dependencies = [
    "fastapi>=0.104",
    "uvicorn[standard]>=0.24",
    "python-multipart>=0.0.6",
    "pydantic>=2.5",
    "pydantic-settings>=2.1",
    "PyYAML>=6.0",
    "structlog>=23.2",
    "numpy>=1.26",
    "opencv-contrib-python>=4.8",
    "torch>=2.1",
    "torchvision>=0.16",
    "mediapipe>=0.10",
    "Pillow>=10.0",
]

[project.optional-dependencies]
train = [
    "tensorboardX>=2.6",
    "tqdm>=4.66",
    "easydict>=1.10",
    "matplotlib>=3.8",
]
dev = [
    "pytest>=7.4",
    "pytest-asyncio>=0.23",
    "httpx>=0.25",
    "ruff>=0.1",
]

[project.scripts]
face-liveness-api = "app.main:run"
face-liveness-turnstile = "app.cli.turnstile_kiosk:main"

[tool.setuptools.packages.find]
include = ["app*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
target-version = "py311"
line-length = 100
